Do not show notice for selected payment method as backup on details
